+++ tap_sftp/singer_encodings/compression.py
