+++ tox.ini
@@ -12,12 +12,12 @@ deps=
     coverage
     pycodestyle
     pytest
-    -rrequirements.txt
+    -r requirements.txt
 whitelist_externals =
     pycodestyle
     coverage
 commands =
-    pycodestyle tap_nicesftp/
+    pycodestyle tap_sftp/
     coverage run -m pytest {posargs}
     coverage html -d tests/codecoverage
 [pycodestyle]
@@ -30,7 +30,7 @@ testpaths =
     tests/tox_tests
 [coverage:run]
 source=
-    tap_nicesftp/
+    tap_sftp/
 omit=
     */.tox/*
     */test/*
